requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
attrs==23.2.0
banal==1.0.6
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
coremltools==7.1
dataset==1.6.2
datasets==2.18.0
dill==0.3.8
filelock==3.13.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.2.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
onnx==1.16.0
onnxruntime==1.17.1
packaging==24.0
pandas==2.2.1
pillow==10.3.0
protobuf==3.20.3
pyaml==23.12.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
six==1.16.0
SQLAlchemy==1.4.52
sympy==1.12
timm==0.9.16
tokenizers==0.15.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.39.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4
